:truck: Change Action trait to View
