+++ src/app.rs
@@ -1,11 +1,11 @@
-use crate::action::Action;
-use crate::buffer::SamplesBuffer;
-use crate::chart::SignalChart;
-use crate::event;
-use crate::file::File;
-use crate::menu::Menu;
-use crate::path;
-use crate::terminal::{self, CrossTerm};
+use crate::dsp::buffer::SamplesBuffer;
+use crate::io::event;
+use crate::io::path;
+use crate::io::terminal::{self, CrossTerm};
+use crate::view::chart::SignalChart;
+use crate::view::file::File;
+use crate::view::menu::Menu;
+use crate::view::View;
 use color_eyre::eyre;
 use crossterm::event::{KeyCode, KeyEvent};
 use rodio::buffer;
@@ -19,7 +19,6 @@ use tui::layout::{Direction, Layout};
 
 /// Main runner for Sampitor application.
 pub struct App {
-    actions: Vec<Box<dyn Action>>,
     menu: Menu,
     samples: SamplesBuffer,
     shutdown: bool,
@@ -27,6 +26,7 @@ pub struct App {
     // If stream is dropped then sound will not reach the speakers.
     _stream: OutputStream,
     terminal: CrossTerm,
+    views: Vec<Box<dyn View>>,
 }
 
 impl App {
@@ -45,21 +45,21 @@ impl App {
         let file = File::try_new(env::current_dir()?)?;
 
         Ok(App {
-            actions: vec![Box::new(chart), Box::new(file)],
             menu: Menu::new(options, String::from("Menu")),
             samples,
             shutdown: false,
             sink,
             _stream: stream,
             terminal: terminal::take()?,
+            views: vec![Box::new(chart), Box::new(file)],
         })
     }
 
     /// Pass keyboard input to current view.
     fn key_event(&mut self, event: KeyEvent) {
         self.menu.key_event(event);
-        let action = &mut self.actions[self.menu.get_state()];
-        action.key_event(event);
+        let view = &mut self.views[self.menu.get_state()];
+        view.key_event(event);
 
         match event.code {
             KeyCode::Char(' ') => self.play(),
@@ -78,7 +78,7 @@ impl App {
 
     /// Render all UI views in terminal screen.
     fn render(&mut self) -> eyre::Result<()> {
-        let action = &mut self.actions[self.menu.get_state()];
+        let view = &mut self.views[self.menu.get_state()];
         let menu = &mut self.menu;
 
         self.terminal.draw(|frame| {
@@ -90,7 +90,7 @@ impl App {
                 .split(size);
 
             menu.render(frame, chunks[0]);
-            action.render(frame, chunks[1]);
+            view.render(frame, chunks[1]);
         })?;
 
         Ok(())
@@ -98,8 +98,8 @@ impl App {
 
     /// Update internal signal state.
     fn process(&mut self) {
-        for action in self.actions.iter_mut() {
-            action.process(&mut self.samples);
+        for view in self.views.iter_mut() {
+            view.process(&mut self.samples);
         }
     }
