+++ src/lib.rs
@@ -1,12 +1,7 @@
-mod action;
 pub mod app;
-mod buffer;
-mod chart;
-mod event;
-mod file;
-mod menu;
-mod path;
-mod terminal;
+mod dsp;
+mod io;
 mod ui;
+mod view;
 
 pub use crate::app::App;
