+++ src/view/mod.rs
@@ -1,0 +1,6 @@
+mod base;
+pub mod chart;
+pub mod file;
+pub mod menu;
+
+pub use base::{CrossFrame, View};
